rename datasets in test folder
