+++ test/datasets/sleep-sample-data.js
@@ -1,0 +1,36 @@
+const sleepData = [
+  {
+    "userID": 1,
+    "date": "2019/06/15",
+    "hoursSlept": 6.1,
+    "sleepQuality": 2.2
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/15",
+    "hoursSlept": 7,
+    "sleepQuality": 4.7
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/15",
+    "hoursSlept": 10.8,
+    "sleepQuality": 4.7
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/15",
+    "hoursSlept": 5.4,
+    "sleepQuality": 3
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/15",
+    "hoursSlept": 4.1,
+    "sleepQuality": 3.6
+  }
+];
+
+if (typeof module !== 'undefined') {
+  module.exports = sleepData;
+}
